[Core] Further workaround of NTLogin

+++ Lagrange.Core/Events/EventArgs/BotLoginEvent.cs
@@ -1,10 +1,10 @@
 namespace Lagrange.Core.Events.EventArgs;
 
-public class BotLoginEvent(byte state, (string, string)? error) : EventBase
+public class BotLoginEvent(int state, (string, string)? error) : EventBase
 {
     public bool Success => State == 0;
     
-    public byte State { get; } = state;
+    public int State { get; } = state;
 
     public (string Tag, string Message)? Error { get; } = error;
